+++ src/main/java/com/tencao/projectbalance/gameObjs/tile/BMPedestalTile.kt
@@ -18,4 +18,4 @@ package com.tencao.projectbalance.gameObjs.tile
 
 import com.tencao.projectbalance.config.ProjectBConfig
 
-class BMPedestalTile: DMPedestalTile(12, ProjectBConfig.tweaks.BMPedestalMax)
+class BMPedestalTile: DMPedestalTile(12, ProjectBConfig.tweaks.BMPedestalMax.toLong())
