+++ src/main/java/com/tencao/projectbalance/gameObjs/tile/RMPedestalTile.kt
@@ -18,4 +18,4 @@ package com.tencao.projectbalance.gameObjs.tile
 
 import com.tencao.projectbalance.config.ProjectBConfig
 
-class RMPedestalTile: DMPedestalTile(8, ProjectBConfig.tweaks.RMPedestalMax)
+class RMPedestalTile: DMPedestalTile(8, ProjectBConfig.tweaks.RMPedestalMax.toLong())
